rotate player instead of gun & more physics integration (bullets, npc)

+++ src/gun.rs
@@ -1,4 +1,5 @@
 use bevy::{prelude::*, window::PrimaryWindow};
+use bevy_rapier2d::prelude::*;
 
 use super::constants::*;
 use super::player;
@@ -16,58 +17,46 @@ pub struct Bullet {
     pub direction: Vec2,
 }
 
+#[allow(clippy::too_many_arguments, clippy::type_complexity)]
 pub fn gun_controls(
     mut commands: Commands,
     window_query: Query<&Window, With<PrimaryWindow>>,
     camera_query: Query<(&Camera, &GlobalTransform), With<player::MainCamera>>,
-    mut player_query: Query<(&Transform, &Children), (With<player::Player>, Without<Gun>)>,
-    mut gun_query: Query<(&mut Gun, &mut Transform), With<Gun>>,
+    mut player_query: Query<(&mut Transform, &Children), (With<player::Player>, Without<Gun>)>,
+    mut gun_query: Query<&mut Gun, With<Gun>>,
     mouse_button_input: Res<Input<MouseButton>>,
+    keyboard_input: Res<Input<KeyCode>>,
     time: Res<Time>,
 ) {
     let window = window_query.get_single().expect("Primary window not found");
     let (camera, camera_transform) = camera_query.get_single().expect("Main camera not found");
 
-    for (player_transform, children) in player_query.iter_mut() {
-        if let Some(gun_entity) = children
-            .iter()
-            .find(|&&entity| gun_query.get_component::<Gun>(entity).is_ok())
-        {
-            if let Ok((mut gun, mut gun_transform)) = gun_query.get_mut(*gun_entity) {
-                gun.shoot_timer -= time.delta_seconds(); // decrement the timer
+    for (mut player_transform, children) in player_query.iter_mut() {
+        if let Some(cursor_position) = window.cursor_position() {
+            if let Some(world_position) = camera.viewport_to_world(camera_transform, cursor_position).map(|ray| ray.origin.truncate()) {
+                let diff = world_position - Vec2::new(player_transform.translation.x, player_transform.translation.y);
+                let angle = diff.y.atan2(diff.x);
 
-                if let Some(cursor_position) = window.cursor_position() {
-                    if let Some(world_position) = camera
-                        .viewport_to_world(camera_transform, cursor_position)
-                        .map(|ray| ray.origin.truncate())
-                    {
-                        let diff = world_position
-                            - Vec2::new(
-                                player_transform.translation.x,
-                                player_transform.translation.y,
-                            );
-                        let angle = diff.y.atan2(diff.x);
+                // rotate player towards cursor
+                player_transform.rotation = Quat::from_axis_angle(Vec3::Z, angle);
 
-                        // Update gun position and rotation
-                        let gun_distance_from_center = PLAYER_SIZE / 2.0 + 10.0; // half player size plus a bit more
-                        let gun_offset = Vec3::new(
-                            gun_distance_from_center * angle.cos(),
-                            gun_distance_from_center * angle.sin(),
-                            0.0,
-                        );
-                        gun_transform.translation = gun_offset;
-                        gun_transform.rotation = Quat::from_axis_angle(Vec3::Z, angle);
+                if let Some(gun_entity) = children.iter().find(|&&entity| gun_query.get_component::<Gun>(entity).is_ok()) {
+                    if let Ok(mut gun) = gun_query.get_mut(*gun_entity) {
+                        gun.shoot_timer -= time.delta_seconds();
 
-                        // Shooting logic
-                        if mouse_button_input.just_pressed(MouseButton::Left)
-                            && gun.shoot_timer <= 0.0
-                        {
-                            gun.shoot_timer = gun.shoot_cooldown; // reset the firing cooldown timer
+                        if (mouse_button_input.just_pressed(MouseButton::Left) || keyboard_input.pressed(KeyCode::Space)) && gun.shoot_timer <= 0.0 {
+                            gun.shoot_timer = gun.shoot_cooldown;
+
+                            // calculate the bullet origin using the player's direction
+                            let bullet_offset = Vec3::new(angle.cos(), angle.sin(), 0.0) * (PLAYER_SIZE / 2.0 + GUN_OFFSET);
+                            let bullet_origin = player_transform.translation + bullet_offset;
+
+                            // spawn bullet
                             spawn_bullet(
                                 &mut commands,
-                                player_transform.translation + gun_transform.translation,
-                                gun_transform.rotation,
-                                diff,
+                                bullet_origin,
+                                player_transform.rotation,
+                                diff.normalize(),
                             );
                         }
                     }
@@ -89,10 +78,16 @@ fn spawn_bullet(commands: &mut Commands, bullet_origin: Vec3, rotation: Quat, di
             },
             sprite: Sprite {
                 color: Color::rgb(0.8, 0.8, 0.0),
-                custom_size: Some(Vec2::new(20.0, 2.0)),
+                custom_size: Some(BULLET_SIZE),
                 ..default()
             },
             ..default()
+        })
+        .insert(RigidBody::Dynamic)
+        .insert(Collider::cuboid(BULLET_SIZE.x / 2., BULLET_SIZE.y  / 2.))
+        .insert(Velocity {
+            linvel: bullet_direction * BULLET_SPEED,
+            angvel: 0.0,
         })
         .insert(Bullet {
             lifetime: BULLET_LIFETIME,
